refactor alert block
