+++ src/blocks/alert/block.json
@@ -1,0 +1,46 @@
+{
+	"name": "alert",
+	"category": "coblocks",
+	"attributes": {
+		"title": {
+			"type": "string",
+			"selector": ".wp-block-coblocks-alert__title"
+		},
+		"value": {
+			"type": "array",
+			"source": "children",
+			"selector": ".wp-block-coblocks-alert__text",
+			"default": []
+		},
+		"backgroundColor": {
+			"type": "string"
+		},
+		"customBackgroundColor": {
+			"type": "string"
+		},
+		"textColor": {
+			"type": "string"
+		},
+		"customTextColor": {
+			"type": "string"
+		},
+		"titleColor": {
+			"type": "string"
+		},
+		"customTitleColor": {
+			"type": "string"
+		},
+		"borderColor": {
+			"type": "string"
+		},
+		"customBorderColor": {
+			"type": "string"
+		},
+		"textAlign": {
+			"type": "string"
+		},
+		"type": {
+			"type": "string"
+		}
+	}
+}
